feat: move and export extendBorders
